+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Fixed
 - fix TypeError from being thrown if the position of definition, highlight range, or rename calculations are from outside the document ([#132](https://github.com/rcjsuen/dockerfile-language-service/issues/132))
+- fix TypeError from being thrown if the position of definition, highlight range, or rename calculations are for tabbed heredocs ([#135](https://github.com/rcjsuen/dockerfile-language-service/issues/135))
 
 ## [0.15.0] - 2025-04-10
 ### Added